تبدیل موجودی بانکی به یک واحد بزرگتر

+++ ViewModels/Areas/Administrator/Bank/IndexViewModel.cs
@@ -25,7 +25,7 @@
         (ResourceType = typeof(Resources.ViewModel.Bank),
             Name = Resources.ViewModel.Strings.BankKeys.Balance)]
         #endregion
-        public int Balance { get; set; }
+        public long Balance { get; set; }
 
         #region InsertDateTime
         [System.ComponentModel.DataAnnotations.Display
